f0.5 Added planting and harvesting
Changed lettuce carrot and tomato.tscn to have Plant.gd as script instead of unique scripts. Recoupled some functions to be in better locations like planting and harvesting functions in Player.gd. Planted seeds are offset from grid, this needs to be fixed. Made water and sun requirements for plants exported variables so they are easy to change and can be unique to each subclass of Plant. Still need to implement the spacial growth requirement.